+++ cmd/description/main.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/google/go-github/v51/github"
 	"github.com/jessevdk/go-flags"
-	"github.com/sashabaranov/go-openai"
 
+	"github.com/ravilushqa/gpt-pullrequest-updater/description"
 	ghClient "github.com/ravilushqa/gpt-pullrequest-updater/github"
 	"github.com/ravilushqa/gpt-pullrequest-updater/jira"
 	oAIClient "github.com/ravilushqa/gpt-pullrequest-updater/openai"
@@ -37,10 +37,6 @@ func main() {
 		os.Exit(0)
 	}
 
-	if opts.Test {
-		fmt.Println("Test mode")
-	}
-
 	if err := run(ctx); err != nil {
 		panic(err)
 	}
@@ -60,25 +56,9 @@ func run(ctx context.Context) error {
 		return fmt.Errorf("error getting commits: %w", err)
 	}
 
-	var sumDiffs int
-	for _, file := range diff.Files {
-		if file.Patch == nil {
-			continue
-		}
-		sumDiffs += len(*file.Patch)
-	}
-
-	var completion string
-	if sumDiffs < 4000 {
-		completion, err = genCompletionOnce(ctx, openAIClient, diff)
-		if err != nil {
-			return fmt.Errorf("error generating completition once: %w", err)
-		}
-	} else {
-		completion, err = genCompletionPerFile(ctx, openAIClient, diff, pr)
-		if err != nil {
-			return fmt.Errorf("error generating completition twice: %w", err)
-		}
+	completion, err := description.GenerateCompletion(ctx, openAIClient, diff, pr)
+	if err != nil {
+		return fmt.Errorf("error generating completion: %w", err)
 	}
 
 	if opts.JiraURL != "" {
@@ -104,72 +84,4 @@ func run(ctx context.Context) error {
 	}
 
 	return nil
-}
-
-func genCompletionOnce(ctx context.Context, client *oAIClient.Client, diff *github.CommitsComparison) (string, error) {
-	fmt.Println("Generating completion once")
-	messages := make([]openai.ChatCompletionMessage, 0, len(diff.Files))
-	messages = append(messages, openai.ChatCompletionMessage{
-		Role:    openai.ChatMessageRoleUser,
-		Content: oAIClient.PromptDescribeChanges,
-	})
-	for _, file := range diff.Files {
-		if file.Patch == nil {
-			continue
-		}
-
-		messages = append(messages, openai.ChatCompletionMessage{
-			Role:    openai.ChatMessageRoleUser,
-			Content: *file.Patch,
-		})
-	}
-
-	fmt.Println("Sending prompt to OpenAI")
-	completion, err := client.ChatCompletion(ctx, messages)
-	if err != nil {
-		return "", fmt.Errorf("error completing prompt: %w", err)
-	}
-
-	return completion, nil
-}
-
-func genCompletionPerFile(ctx context.Context, client *oAIClient.Client, diff *github.CommitsComparison, pr *github.PullRequest) (string, error) {
-	fmt.Println("Generating completion per file")
-	OverallDescribeCompletion := fmt.Sprintf("Pull request title: %s, body: %s\n\n", pr.GetTitle(), pr.GetBody())
-
-	for i, file := range diff.Files {
-		if file.Patch == nil {
-			continue
-		}
-		prompt := fmt.Sprintf(oAIClient.PromptDescribeChanges, *file.Patch)
-
-		if len(prompt) > 4096 {
-			prompt = fmt.Sprintf("%s...", prompt[:4093])
-		}
-
-		fmt.Printf("Sending prompt to OpenAI for file %d/%d\n", i+1, len(diff.Files))
-		completion, err := client.ChatCompletion(ctx, []openai.ChatCompletionMessage{
-			{
-				Role:    openai.ChatMessageRoleUser,
-				Content: prompt,
-			},
-		})
-		if err != nil {
-			return "", fmt.Errorf("error getting review: %w", err)
-		}
-		OverallDescribeCompletion += fmt.Sprintf("File: %s \nDescription: %s \n\n", file.GetFilename(), completion)
-	}
-
-	fmt.Println("Sending final prompt to OpenAI")
-	overallCompletion, err := client.ChatCompletion(ctx, []openai.ChatCompletionMessage{
-		{
-			Role:    openai.ChatMessageRoleUser,
-			Content: fmt.Sprintf(oAIClient.PromptOverallDescribe, OverallDescribeCompletion),
-		},
-	})
-	if err != nil {
-		return "", fmt.Errorf("error getting overall review: %w", err)
-	}
-
-	return overallCompletion, nil
 }
