+++ openai/openai.go
@@ -12,10 +12,11 @@ import (
 //go:embed prompts/review
 var PromptReview string
 
-const (
-	PromptDescribeChanges = "Below is the code patch, Generate a GitHub pull request description based on the following comments without basic prefix\n%s\n"
-	PromptOverallDescribe = "Below comments are generated by AI, Generate a GitHub pull request description based on the following comments without basic prefix in markdown format with ### Description and ### Changes blocks:\n%s\n"
-)
+//go:embed prompts/describe_changes
+var PromptDescribeChanges string
+
+//go:embed prompts/describe_overall
+var PromptDescribeOverall string
 
 type Client struct {
 	client *openai.Client
