+++ openai/prompts/describe_changes
@@ -1,0 +1,2 @@
+Below is the code patch, Generate a GitHub pull request description based on the following comments without basic prefix
+%s
