+++ openai/prompts/describe_overall
@@ -1,0 +1,6 @@
+Below comments are generated by AI
+Generate a GitHub pull request description based on the following comments without basic prefix in markdown format.
+Response should contain ### Description and ### Changes blocks
+
+PR comments:
+%s
